#16 List SportComplexDetailedDto with pagination
