+++ backend/initdb/02-random-data.sql
@@ -13,6 +13,26 @@ VALUES
     ('Green Park Wellness Center', 'Comprehensive wellness center offering various health and fitness services.', 'http://www.greenparkwellness.com', 'TRAWA', 'CROSSFIT', 54.356805, 18.658322, 2, TRUE, 'green_park_wellness.jpg', '+48111222334'),
     ('Mountain Peak Ski Lodge', 'Exclusive ski lodge with beginner to expert slopes and luxurious amenities.', 'http://www.mountainpeakski.com', 'LÓD', 'FITNESS', 54.341546, 18.649502, 3, FALSE, 'mountain_peak_ski.jpg', '+48123444556'),
     ('Sunset Yoga Retreat', 'Serene yoga retreat offering a variety of classes and wellness programs.', 'http://www.sunsetyoga.com', 'DREWNIANA_PODŁOGA', 'KOSZYKÓWKA', 54.360686, 18.635629, 4, FALSE, 'sunset_yoga.jpg', '+48123455667'),
+    ('River Edge Tennis Academy', 'State-of-the-art tennis academy with professional training and facilities.', 'http://www.riveredgetennis.com', 'ASFALT', 'LEKKOATLETYKA', 54.348629, 18.676230, 5, TRUE, 'river_edge_tennis.jpg', '+48123466778'),
+    ('New Olympic Fitness Center', 'An advanced fitness center with top-tier equipment and diverse training options.', 'http://www.newolympicfitness.com', 'ASFALT', 'BIEGANIE', 54.352025, 18.646638, 1, FALSE, 'new_olympic_fitness.jpg', '+48123456789'),
+    ('Green Park Wellness Center', 'Comprehensive wellness center offering various health and fitness services.', 'http://www.greenparkwellness.com', 'TRAWA', 'CROSSFIT', 54.356805, 18.658322, 2, TRUE, 'green_park_wellness.jpg', '+48111222334'),
+    ('Mountain Peak Ski Lodge', 'Exclusive ski lodge with beginner to expert slopes and luxurious amenities.', 'http://www.mountainpeakski.com', 'LÓD', 'FITNESS', 54.341546, 18.649502, 3, FALSE, 'mountain_peak_ski.jpg', '+48123444556'),
+    ('Sunset Yoga Retreat', 'Serene yoga retreat offering a variety of classes and wellness programs.', 'http://www.sunsetyoga.com', 'DREWNIANA_PODŁOGA', 'KOSZYKÓWKA', 54.360686, 18.635629, 4, FALSE, 'sunset_yoga.jpg', '+48123455667'),
+    ('River Edge Tennis Academy', 'State-of-the-art tennis academy with professional training and facilities.', 'http://www.riveredgetennis.com', 'ASFALT', 'LEKKOATLETYKA', 54.348629, 18.676230, 5, TRUE, 'river_edge_tennis.jpg', '+48123466778'),
+    ('New Olympic Fitness Center', 'An advanced fitness center with top-tier equipment and diverse training options.', 'http://www.newolympicfitness.com', 'ASFALT', 'BIEGANIE', 54.352025, 18.646638, 1, FALSE, 'new_olympic_fitness.jpg', '+48123456789'),
+    ('Green Park Wellness Center', 'Comprehensive wellness center offering various health and fitness services.', 'http://www.greenparkwellness.com', 'TRAWA', 'CROSSFIT', 54.356805, 18.658322, 2, TRUE, 'green_park_wellness.jpg', '+48111222334'),
+    ('Mountain Peak Ski Lodge', 'Exclusive ski lodge with beginner to expert slopes and luxurious amenities.', 'http://www.mountainpeakski.com', 'LÓD', 'FITNESS', 54.341546, 18.649502, 3, FALSE, 'mountain_peak_ski.jpg', '+48123444556'),
+    ('Sunset Yoga Retreat', 'Serene yoga retreat offering a variety of classes and wellness programs.', 'http://www.sunsetyoga.com', 'DREWNIANA_PODŁOGA', 'KOSZYKÓWKA', 54.360686, 18.635629, 4, FALSE, 'sunset_yoga.jpg', '+48123455667'),
+    ('River Edge Tennis Academy', 'State-of-the-art tennis academy with professional training and facilities.', 'http://www.riveredgetennis.com', 'ASFALT', 'LEKKOATLETYKA', 54.348629, 18.676230, 5, TRUE, 'river_edge_tennis.jpg', '+48123466778'),
+    ('New Olympic Fitness Center', 'An advanced fitness center with top-tier equipment and diverse training options.', 'http://www.newolympicfitness.com', 'ASFALT', 'BIEGANIE', 54.352025, 18.646638, 1, FALSE, 'new_olympic_fitness.jpg', '+48123456789'),
+    ('Green Park Wellness Center', 'Comprehensive wellness center offering various health and fitness services.', 'http://www.greenparkwellness.com', 'TRAWA', 'CROSSFIT', 54.356805, 18.658322, 2, TRUE, 'green_park_wellness.jpg', '+48111222334'),
+    ('Mountain Peak Ski Lodge', 'Exclusive ski lodge with beginner to expert slopes and luxurious amenities.', 'http://www.mountainpeakski.com', 'LÓD', 'FITNESS', 54.341546, 18.649502, 3, FALSE, 'mountain_peak_ski.jpg', '+48123444556'),
+    ('Sunset Yoga Retreat', 'Serene yoga retreat offering a variety of classes and wellness programs.', 'http://www.sunsetyoga.com', 'DREWNIANA_PODŁOGA', 'KOSZYKÓWKA', 54.360686, 18.635629, 4, FALSE, 'sunset_yoga.jpg', '+48123455667'),
+    ('River Edge Tennis Academy', 'State-of-the-art tennis academy with professional training and facilities.', 'http://www.riveredgetennis.com', 'ASFALT', 'LEKKOATLETYKA', 54.348629, 18.676230, 5, TRUE, 'river_edge_tennis.jpg', '+48123466778'),
+    ('New Olympic Fitness Center', 'An advanced fitness center with top-tier equipment and diverse training options.', 'http://www.newolympicfitness.com', 'ASFALT', 'BIEGANIE', 54.352025, 18.646638, 1, FALSE, 'new_olympic_fitness.jpg', '+48123456789'),
+    ('Green Park Wellness Center', 'Comprehensive wellness center offering various health and fitness services.', 'http://www.greenparkwellness.com', 'TRAWA', 'CROSSFIT', 54.356805, 18.658322, 2, TRUE, 'green_park_wellness.jpg', '+48111222334'),
+    ('Mountain Peak Ski Lodge', 'Exclusive ski lodge with beginner to expert slopes and luxurious amenities.', 'http://www.mountainpeakski.com', 'LÓD', 'FITNESS', 54.341546, 18.649502, 3, FALSE, 'mountain_peak_ski.jpg', '+48123444556'),
+    ('Sunset Yoga Retreat', 'Serene yoga retreat offering a variety of classes and wellness programs.', 'http://www.sunsetyoga.com', 'DREWNIANA_PODŁOGA', 'KOSZYKÓWKA', 54.360686, 18.635629, 4, FALSE, 'sunset_yoga.jpg', '+48123455667'),
     ('River Edge Tennis Academy', 'State-of-the-art tennis academy with professional training and facilities.', 'http://www.riveredgetennis.com', 'ASFALT', 'LEKKOATLETYKA', 54.348629, 18.676230, 5, TRUE, 'river_edge_tennis.jpg', '+48123466778');
 
 INSERT INTO opening_hours (day_of_week, opening_time, closing_time, sport_complex_id)
